requirements.txt
django>=3.1.1
psycopg2==2.8.5
djangorestframework==3.11
pygments==2.6
django-cors-headers==3.4.0
# daphne==2.5.0
django-q>=1.3.4
pypandoc==1.5
requests>=2.24.0
antlr4-python3-runtime==4.8
six==1.15.0
